requirements.txt
Django==5.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-froala-editor==4.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
whitenoise==6.6.0
Pillow==10.2.0
pandas==2.2.1
openpyxl==3.1.2
gunicorn==21.2.0
